Added the mail option to gisty
Added a simple mailto in the application to allow for gists to be easily shared thru email.  Opted for this method over a backend server (which would have allow for html email formatting) because it went with gisty's no server theme.  It prevents the need for backend servers that could possible scrape data from the end users - at least it's not on gisty's side :)